+++ android/src/main/java/it/remedia/flutter_facebook_app_links/FlutterFacebookAppLinksPlugin.java
@@ -3,7 +3,8 @@ package it.remedia.flutter_facebook_app_links;
 import android.app.Activity;
 import android.content.Context;
 import android.os.Handler;
-//import android.util.Log;
+import android.util.Log;
+import androidx.annotation.NonNull;
 
 import com.facebook.applinks.AppLinkData;
 import com.facebook.FacebookSdk;
@@ -24,21 +25,24 @@ public class FlutterFacebookAppLinksPlugin implements FlutterPlugin, MethodCallH
 
   private Context mContext;
   private Activity mActivity;
-  private MethodChannel channel;
+  private String deeplinkUrl = "";
+  private MethodChannel methodChannel;
 
+  private static String TAG = "FlutterFacebookAppLinksPlugin";
   private static final String CHANNEL = "plugins.remedia.it/flutter_facebook_app_links";
 
   @Override
-  public void onAttachedToEngine(FlutterPluginBinding binding) {
-    channel = new MethodChannel(binding.getBinaryMessenger(), CHANNEL);
-    channel.setMethodCallHandler(this);
+  public void onAttachedToEngine(@NonNull FlutterPluginBinding binding) {
+    Log.d(TAG, "onAttachedToEngine...");
+    methodChannel = new MethodChannel(binding.getBinaryMessenger(), CHANNEL);
+    methodChannel.setMethodCallHandler(this);
     mContext = binding.getApplicationContext();
   }
 
   @Override
-  public void onDetachedFromEngine(FlutterPluginBinding binding) {
-    channel.setMethodCallHandler(null);
-    channel = null;
+  public void onDetachedFromEngine(@NonNull FlutterPluginBinding binding) {
+    methodChannel.setMethodCallHandler(null);
+    methodChannel = null;
     mContext = null;
   }
 
@@ -69,7 +73,7 @@ public class FlutterFacebookAppLinksPlugin implements FlutterPlugin, MethodCallH
     } else if(call.method.equals("initFBLinks")){
       initFBLinks(result);
     } else if (call.method.equals("getDeepLinkUrl")) {
-      getDeepLinkUrl(result);
+      result.success(deeplinkUrl);
     } else if (call.method.equals("consentProvided")) {
       FacebookSdk.setAutoLogAppEventsEnabled(true);
       FacebookSdk.setAutoInitEnabled(true);
@@ -80,68 +84,30 @@ public class FlutterFacebookAppLinksPlugin implements FlutterPlugin, MethodCallH
       FacebookSdk.setAutoInitEnabled(false);
       FacebookSdk.fullyInitialize();
       result.success("");
+    } else if(call.method.equals("activateApp")){
+      activateSDK();
+      result.success(true);
     } else {
       result.notImplemented();
     }
   }
 
-  private void getDeepLinkUrl(Result result) {
-    //Log.d("FB_APP_LINKS", "Facebook App Links getDeepLinkUrl called");
-
-    final Result resultDelegate = result;
-    // Get a handler that can be used to post to the main thread
-    final Handler mainHandler = new Handler(mContext.getMainLooper());
-
-    // Get user consent
+  private void activateSDK(){
+    FacebookSdk.setAutoLogAppEventsEnabled(false);
+    FacebookSdk.setAutoInitEnabled(true);
     FacebookSdk.fullyInitialize();
-    AppLinkData.fetchDeferredAppLinkData(mContext,
-      new AppLinkData.CompletionHandler() {
-        @Override
-        public void onDeferredAppLinkDataFetched(AppLinkData appLinkData) {
-          // Process app link data
-          if(appLinkData!=null && appLinkData.getTargetUri()!=null){
-            //Log.d("FB_APP_LINKS", "Deep Link URL Received: " + appLinkData.getTargetUri().toString());
-            
-            Runnable myRunnable = new Runnable() {
-              @Override
-              public void run() {
-                if(resultDelegate!=null)
-                  resultDelegate.success(appLinkData.getTargetUri().toString());
-              }
-            };
-
-            mainHandler.post(myRunnable);
-
-          }else{
-            //Log.d("FB_APP_LINKS", "Deep Link URL Received: null link");
-
-            Runnable myRunnable = new Runnable() {
-              @Override
-              public void run() {
-                if(resultDelegate!=null)
-                  resultDelegate.success("");
-              }
-            };
-
-            mainHandler.post(myRunnable);
-
-          }
-
-        }
-      }
-    );
   }
 
   private void initFBLinks(Result result) {
     //Log.d("FB_APP_LINKS", "Facebook App Links initialized");
 
-    final Map<String, String> data = new HashMap<>();
     final Result resultDelegate = result;
     // Get a handler that can be used to post to the main thread
     final Handler mainHandler = new Handler(mContext.getMainLooper());
 
     // Get user consent
-    FacebookSdk.fullyInitialize();
+    activateSDK();
+
     AppLinkData.fetchDeferredAppLinkData(mContext,
       new AppLinkData.CompletionHandler() {
         @Override
@@ -151,20 +117,16 @@ public class FlutterFacebookAppLinksPlugin implements FlutterPlugin, MethodCallH
 
             if(appLinkData.getTargetUri()!=null){
               //Log.d("FB_APP_LINKS", "Deferred Deeplink Received: " + appLinkData.getTargetUri().toString());
-              data.put("deeplink", appLinkData.getTargetUri().toString());
+              deeplinkUrl = appLinkData.getTargetUri().toString();
             }
 
             //Log.d("FB_APP_LINKS", "Deferred Deeplink Received: " + appLinkData.getPromotionCode());
-            if(appLinkData.getPromotionCode()!=null)
-              data.put("promotionalCode", appLinkData.getPromotionCode());
-            else
-              data.put("promotionalCode", "");
 
             Runnable myRunnable = new Runnable() {
               @Override
               public void run() {
                 if(resultDelegate!=null)
-                  resultDelegate.success(data);
+                  resultDelegate.success(deeplinkUrl);
               }
             };
 
@@ -177,7 +139,7 @@ public class FlutterFacebookAppLinksPlugin implements FlutterPlugin, MethodCallH
               @Override
               public void run() {
                 if(resultDelegate!=null)
-                  resultDelegate.success(null);
+                  resultDelegate.success(deeplinkUrl);
               }
             };
 
@@ -189,6 +151,5 @@ public class FlutterFacebookAppLinksPlugin implements FlutterPlugin, MethodCallH
       }
     );
   }
-
 
 }