+++ ios/Classes/SwiftFlutterFacebookAppLinksPlugin.swift
@@ -2,18 +2,48 @@ import FBSDKCoreKit
 import Flutter
 import UIKit
 
+
 public class SwiftFlutterFacebookAppLinksPlugin: NSObject, FlutterPlugin {
-  // fileprivate var resulter: FlutterResult? = nil
+
+  var deepLinkUrl:String = ""
 
   public static func register(with registrar: FlutterPluginRegistrar) {
-    let channel = FlutterMethodChannel(name: "plugins.remedia.it/flutter_facebook_app_links", binaryMessenger: registrar.messenger())
-    let instance = SwiftFlutterFacebookAppLinksPlugin()
 
+    let instance = SwiftFlutterFacebookAppLinksPlugin()
+    let channel = FlutterMethodChannel(name: "plugins.remedia.it/flutter_facebook_app_links", binaryMessenger: registrar.messenger())
+    
     // Get user consent
     print("FB APP LINK registering plugin")
-    ApplicationDelegate.shared.initializeSDK()
-
+    
+    instance.initializeSDK()
+    
     registrar.addMethodCallDelegate(instance, channel: channel)
+    registrar.addApplicationDelegate(instance)
+  }
+
+  public func detachFromEngine(for registrar: FlutterPluginRegistrar) {
+    // detach
+  }
+
+  public func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [AnyHashable : Any] = [:]) -> Bool {
+
+      Settings.shared.isAdvertiserTrackingEnabled = false
+      let launchOptionsForFacebook = launchOptions as? [UIApplication.LaunchOptionsKey: Any]
+      ApplicationDelegate.shared.application(
+          application,
+          didFinishLaunchingWithOptions:
+              launchOptionsForFacebook
+      )
+      AppLinkUtility.fetchDeferredAppLink{ (url, error) in
+          if let error = error{
+              print("Error %a", error)
+          }
+          if let url = url {
+              self.deepLinkUrl = url.absoluteString
+              // self.sendMessageToStream(link: self.deepLinkUrl)
+          }
+      }
+      return true
   }
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
@@ -33,50 +63,21 @@ public class SwiftFlutterFacebookAppLinksPlugin: NSObject, FlutterPlugin {
       handleFBAppLinks(call, result: result)
     case "getDeepLinkUrl":
       print("FB APP LINK getDeepLinkUrl called")
-      handleGetDeepLinkUrl(call, result: result)
+      result(deepLinkUrl)
+    case "activateApp":
+      AppEvents.shared.activateApp()
+      result(true)
     default:
       result(FlutterMethodNotImplemented)
     }
   }
 
-  private func handleGetPlatformVersion(_: FlutterMethodCall, result: @escaping FlutterResult) {
+  private func handleGetPlatformVersion(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     result("iOS " + UIDevice.current.systemVersion)
   }
 
-  private func handleFBAppLinks(_: FlutterMethodCall, result: @escaping FlutterResult) {
+  private func handleFBAppLinks(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     print("FB APP LINKS Starting ")
-
-    AppLinkUtility.fetchDeferredAppLink { url, error in
-      if let error = error {
-        print("Received error while fetching deferred app link %@", error)
-        result(nil)
-      }
-
-      if let url = url {
-        print("FB APP LINKS getting url: ", String(url.absoluteString))
-
-        var mapData: [String: String?] = ["deeplink": url.absoluteString, "promotionalCode": nil]
-
-        if let code = AppLinkUtility.appInvitePromotionCode(from: url) {
-          print("promotional code " + String(code))
-          mapData["promotionalCode"] = code
-        } else { // nil
-        }
-
-        if #available(iOS 10, *) {
-          result(mapData)
-        } else {
-          result(mapData)
-        }
-      } else {
-        // no deep link received
-        result(nil)
-      }
-    }
-  }
-
-  private func handleGetDeepLinkUrl(_: FlutterMethodCall, result: @escaping FlutterResult) {
-    print("FB APP LINKS getDeepLinkUrl Starting ")
 
     AppLinkUtility.fetchDeferredAppLink { url, error in
       if let error = error {
@@ -85,12 +86,16 @@ public class SwiftFlutterFacebookAppLinksPlugin: NSObject, FlutterPlugin {
       }
 
       if let url = url {
-        print("FB APP LINKS getDeepLinkUrl getting url: ", String(url.absoluteString))
+        print("FB APP LINKS getting url: ", String(url.absoluteString))
         result(url.absoluteString)
       } else {
         // no deep link received
         result("")
       }
     }
+  }
+
+  public func initializeSDK() {
+    ApplicationDelegate.shared.initializeSDK()
   }
 }