+++ lib/flutter_facebook_app_links.dart
@@ -1,38 +1,33 @@
 import 'dart:async';
-import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter/services.dart';
 
 class FlutterFacebookAppLinks {
   static const MethodChannel _channel = const MethodChannel("plugins.remedia.it/flutter_facebook_app_links");
 
-  static Future<String> get platformVersion async {
-    final String version = await _channel.invokeMethod('getPlatformVersion');
+  static Future<String?> get platformVersion async {
+    final String? version = await _channel.invokeMethod('getPlatformVersion');
     return version;
   }
 
-  static Future<dynamic> initFBLinks() async {
+  static Future<String> initFBLinks() async {
     try {
       var data = await _channel.invokeMethod('initFBLinks');
-
-      if (data == null) return null;
-
-      final Map<String, String> result = new Map.from(data);
-      return result;
+      debugPrint('Deferred FB Link: $data');
+      return data ?? '';
     } catch (e) {
-      debugPrint("Error retrieving deferred deep link: $e");
-
-      return null;
+      debugPrint("Error initializing FlutterFacebookAppLinks: $e");
+      return '';
     }
   }
 
   static Future<String> getDeepLink() async {
     try {
       var data = await _channel.invokeMethod('getDeepLinkUrl');
-      print('Deferred FB Link: $data');
+      debugPrint('Deferred FB Link: $data');
       return data ?? '';
     } catch (e) {
       debugPrint("Error retrieving deferred deep link: $e");
-
       return '';
     }
   }
@@ -43,5 +38,14 @@ class FlutterFacebookAppLinks {
 
   static Future<dynamic> consentRevoked() {
     return _channel.invokeMethod('consentRevoked');
+  }
+
+  static Future<void> activateSDK() async {
+    try {
+      debugPrint('Activating SDK');
+      await _channel.invokeMethod('activateApp');
+    } catch (e) {
+      debugPrint("Error activating SDK: $e");
+    }
   }
 }